Move BFT time estimate to separate module
So it won't get in the way. It conatins too many unused but potencially useful
code
